add fragment_product_info & lattie

+++ app/src/main/java/com/example/onlineshop/data/model/ProductInfo.kt
@@ -5,7 +5,7 @@ data class ProductInfo(
     val ratingCount: Int,
     val description: String,
     val averageRating: String,
-    val relatedList: List<Int>,
+    val relatedList: List<Long>,
     val imagesUrl: List<String>,
     val dimensions: List<String>,
     val attributes: List<ProductAttribute>
